+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 * Added `compas_fab.planning` that contains `Action` classes and `SceneState`
-* Added `RoboticAction`, `CartesianMotion`, `FreeMotion`, `OpenGripper` and `CloseGripper`.
+* Added `RoboticAction`, `CartesianMotion`, `FreeMotion`, `OpenGripper`, `CloseGripper`, and `ManuallyMoveWorkpiece` actions.
 * Added `SceneState` as container for `WorkpieceState`, `ToolState` and `RobotState`.
 
 ### Changed